whoshome/requirements.txt
pydantic==2.10.0
ecs-logging==2.2.0
flask==3.1.0
Flask-Pydantic==0.12.0
requests==2.32.3
pyjwt==2.10.1